fixup! mingw: include the full version information in the resources
The previous attempts to fix this were all piecemeal, let's just try to
ignore everything that is not a digit. Yes, sure, v2.10.0-rc2.windows.1
gets the bogus patchlevel 2 (because of -rc2), but there is not really
anything we can do about that, given that the .rc only allows to use
four numbers to describe the version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>