mingw: add experimental feature to redirect standard handles
On Windows, file handles need to be marked inheritable when they need to
be used as standard handles for a newly spawned process. The problem with
that, of course, is that the "inheritable" flag is global and therefore
can wreak havoc with highly multi-threaded applications.

Let's introduce a set of environment variables (GIT_REDIRECT_STDIN and
friends) that point to files, or even better, named pipes and that are
used by the spawned Git process. This helps work around above-mentioned
issue.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>