config: validate ownership of C:\ProgramData\Git\config before using it
When the VirtualStore feature is in effect, it is safe to let random
users write into C:\ProgramData because other users won't see those
files. This seemed to be the case when we introduced support for
C:\ProgramData\Git\config.

However, when that feature is not in effect (which seems to be the case
in newer Windows 10 versions), we'd rather not use those files unless
they come from a trusted source, such as an administrator.

This change imitates the strategy chosen by PowerShell's native OpenSSH
port to Windows: if a system file is owned neither by an administrator,
a system account, or the current user, it is ignored.

Much as we'd like, we cannot show a pretty advice here, as the advice
machinery depends on the config to already have been read, and at this
point, we have not even started doing so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>