fixup! Windows: add support for a Windows-wide configuration
In the previous commit, we moved the location of the system config to
`C:\Program Files\Git\etc\gitconfig`, to give users a reliable location
for the system config.

Previously, we had introduced `C:\ProgramData\Git\config` as such a
location.

However, the purpose of `C:\ProgramData\Git\config` was always for the
installed Git for Windows to share its configured options. Every upgrade
will modify the settings in that file according to what the user
specified when running Git for Windows' installer. So it is not really a
file whose ownership is shared between all Git implementations, contrary
to what the location suggests.

Also, it was totally confusing to many users that there was a "system"
config and then yet another "sort of system" config that did not even
have a corresponding `git config` option such as `--system`.

Further, Portable Git should probably never look at
`C:\ProgramData\Git\config` to begin with: this makes it rather
non-Portable ;-)

Finally, support for `C:\ProgramData\Git\config` never really caught on:
only libgit2 implemented it, but even JGit (which might be used by more
users than libgit2-based applications by grace of backing Android
Studio) did not.

Therefore, let's finally admit that it was a mistake to introduce
`C:\ProgramData\Git\config` and remove support for it.

This reverts commit bcbbb4f89f57cdd6d2260d28b274e9e6334d7ed3.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>