mingw: allow building with current mingw-w64
Current mingw-w64 defaults to compatibility with Windows Vista or later.
Our current source code does not build with that default, as it still
has code that does not expect symbols introduced in Windows Vista, such
as the pthread compatibility functions or the IPv6 support.

Let's force the headers to Windows XP compatibility, allowing the source
code to be compiled again in a current Git for Windows SDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>