ci(windows): run the tests in parallel
Not only in parallel on one machine, but distribute the load on 10
different jobs. That makes the over-all wait time much more bearable
than having to wait for a full hour every time we want to gain
confidence in a certain set of patches in preparation for a new MinGit
backport.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>